fix controller etc
