+++ app/controllers/users/registrations_controller.rb
@@ -1,49 +1,49 @@
 # frozen_string_literal: true
 
 class Users::RegistrationsController < Devise::RegistrationsController
-  # before_action :configure_sign_up_params, only: [:create]
-  # before_action :configure_account_update_params, only: [:update]
-  before_action :authenticate_scope!, except: %i[new create]
+  def after_update_path_for(_resource)
+    edit_user_registration_path # アカウント情報更新後に遷移するpathを設定
+  end
 
   # GET /resource/sign_up
-  def new
-    build_resource
-    yield resource if block_given?
-    respond_with resource
-  end
+  # def new
+  #   build_resource
+  #   yield resource if block_given?
+  #   respond_with resource
+  # end
 
-  def index
-    @users = User.order(:id).page(params[:page])
-    render :index
-  end
+  # def index
+  #   @users = User.order(:id).page(params[:page])
+  #   render :index
+  # end
 
-  def show
-    @user = User.find(params[:id])
-    render :show
-  end
+  # def show
+  #   @user = User.find(params[:id])
+  #   render :show
+  # end
 
   # POST /resource
-  def create
-    build_resource(sign_up_params)
+  # def create
+  #   build_resource(sign_up_params)
 
-    resource.save
-    yield resource if block_given?
-    if resource.persisted?
-      if resource.active_for_authentication?
-        set_flash_message! :notice, :signed_up
-        sign_up(resource_name, resource)
-        respond_with resource, location: after_sign_up_path_for(resource)
-      else
-        set_flash_message! :notice, :"signed_up_but_#{resource.inactive_message}"
-        expire_data_after_sign_in!
-        respond_with resource, location: after_inactive_sign_up_path_for(resource)
-      end
-    else
-      clean_up_passwords resource
-      set_minimum_password_length
-      respond_with resource
-    end
-  end
+  #   resource.save
+  #   yield resource if block_given?
+  #   if resource.persisted?
+  #     if resource.active_for_authentication?
+  #       set_flash_message! :notice, :signed_up
+  #       sign_up(resource_name, resource)
+  #       respond_with resource, location: after_sign_up_path_for(resource)
+  #     else
+  #       set_flash_message! :notice, :"signed_up_but_#{resource.inactive_message}"
+  #       expire_data_after_sign_in!
+  #       respond_with resource, location: after_inactive_sign_up_path_for(resource)
+  #     end
+  #   else
+  #     clean_up_passwords resource
+  #     set_minimum_password_length
+  #     respond_with resource
+  #   end
+  # end
 
   # GET /resource/edit
   # def edit
@@ -68,12 +68,6 @@ class Users::RegistrationsController < Devise::RegistrationsController
   # def cancel
   #   super
   # end
-
-  protected
-
-  def after_update_path_for(resource)
-    user_path(resource)
-  end
 
   # If you have extra params to permit, append them to the sanitizer.
   # def configure_sign_up_params
