+++ app/controllers/users_controller.rb
@@ -1,0 +1,13 @@
+# frozen_string_literal: true
+
+class UsersController < ApplicationController
+  def index
+    @users = User.order(:id).page(params[:page])
+    render :index
+  end
+
+  def show
+    @user = User.find(params[:id])
+    render :show
+  end
+end
