+++ app/helpers/users_helper.rb
@@ -1,0 +1,15 @@
+# frozen_string_literal: true
+
+module UsersHelper
+  def resource
+    @resource ||= User.new
+  end
+
+  def resource_name
+    :user
+  end
+
+  def devise_mapping
+    @devise_mapping ||= Devise.mappings[:user]
+  end
+end
