+++ app/views/books/edit.html.erb
@@ -2,5 +2,5 @@
 
 <%= render 'form', book: @book %>
 
-<%= link_to t(:show), @book %> |
-<%= link_to t("Back"), books_path %>
+<%= link_to t('views.common.show'), @book %> |
+<%= link_to t('views.common.back'), books_path %>
