+++ app/views/books/new.html.erb
@@ -2,4 +2,4 @@
 
 <%= render 'form', book: @book %>
 
-<%= link_to t("Back"), books_path %>
+<%= link_to t('views.common.back'), books_path %>
