+++ app/views/books/show.html.erb
@@ -18,6 +18,6 @@
   <%= image_tag(@book.picture_url) if @book.picture.present? %>
 </p>
 
-<%= link_to t("Edit"), :edit_book %>
-<%= link_to t("Back"), :back %>
+<%= link_to t('views.common.edit'), :edit_book %>
+<%= link_to t('views.common.back'), :back %>
