--- app/views/devise/mailer/confirmation_instructions.html.erb
@@ -1,5 +1,0 @@
-<p><%= t('.greeting', recipient: @email) %></p>
-
-<p><%= t('.instruction') %></p>
-
-<p><%= link_to t('.action'), confirmation_url(@resource, confirmation_token: @token) %></p>
