--- app/views/devise/mailer/email_changed.html.erb
@@ -1,7 +1,0 @@
-<p><%= t('.greeting', recipient: @email) %></p>
-
-<% if @resource.try(:unconfirmed_email?) %>
-  <p><%= t('.message_unconfirmed', email: @resource.unconfirmed_email) %></p>
-<% else %>
-  <p><%= t('.message', email: @resource.email) %></p>
-<% end %>
