--- app/views/devise/mailer/password_change.html.erb
@@ -1,3 +1,0 @@
-<p><%= t('.greeting', recipient: @resource.email) %></p>
-
-<p><%= t('.message') %></p>
