--- app/views/devise/mailer/reset_password_instructions.html.erb
@@ -1,8 +1,0 @@
-<p><%= t('.greeting', recipient: @resource.email) %></p>
-
-<p><%= t('.instruction') %></p>
-
-<p><%= link_to t('.action'), edit_password_url(@resource, reset_password_token: @token) %></p>
-
-<p><%= t('.instruction_2') %></p>
-<p><%= t('.instruction_3') %></p>
