--- app/views/devise/unlocks/new.html.erb
@@ -1,16 +1,0 @@
-<h2><%= t('.resend_unlock_instructions') %></h2>
-
-<%= form_for(resource, as: resource_name, url: unlock_path(resource_name), html: { method: :post }) do |f| %>
-  <%= render "devise/shared/error_messages", resource: resource %>
-
-  <div class="field">
-    <%= f.label :email %><br />
-    <%= f.email_field :email, autofocus: true, autocomplete: "email" %>
-  </div>
-
-  <div class="actions">
-    <%= f.submit t('.resend_unlock_instructions') %>
-  </div>
-<% end %>
-
-<%= render "devise/shared/links" %>
