+++ app/views/layouts/application.html.erb
@@ -20,13 +20,13 @@
     <p class="navbar-text pull-right">
       <% if user_signed_in? %>
         <%= t('logged_in_as', login_email: current_user.email) %>
-        <%= link_to t('Edit_profile'), edit_user_registration_path, class: 'navbar-link' %> |
-        <%= link_to t('Book_index'), books_path, class: 'navbar-link' %> |
-        <%= link_to t('User_index'), users_path, class: 'navbar-link'  %> |
-        <%= link_to t('Logout'), destroy_user_session_path, method: :delete, class: 'navbar-link'  %>
+        <%= link_to t('edit_profile'), edit_user_registration_path, class: 'navbar-link' %> |
+        <%= link_to t('book_index'), books_path, class: 'navbar-link' %> |
+        <%= link_to t('user_index'), users_path, class: 'navbar-link'  %> |
+        <%= link_to t('logout'), destroy_user_session_path, method: :delete, class: 'navbar-link'  %>
       <% else %>
-        <%= link_to t("Sign_up"), new_user_registration_path, class: 'navbar-link'  %> |
-        <%= link_to t("Log_in"), new_user_session_path, class: 'navbar-link'  %> 
+        <%= link_to t("sign_up"), new_user_registration_path, class: 'navbar-link'  %> |
+        <%= link_to t("log_in"), new_user_session_path, class: 'navbar-link'  %> 
 
       <% end %>
     </p>
