+++ app/views/users/index.html.erb
@@ -1,6 +1,6 @@
 <p id="notice"><%= notice %></p>
 
-<h1><%= t('User_index') %></h1>
+<h1><%= t('user_index') %></h1>
 
 <table>
   <thead>
