+++ app/views/users/show.html.erb
@@ -18,4 +18,4 @@
   <%= @user.self_introduction %>
 </p>
 
-<%= link_to t('Back'), users_path %>
+<%= link_to t('views.common.back'), users_path %>
