+++ config/locales/ja.yml
@@ -25,37 +25,59 @@ ja:
         remember_me: ログインを記憶する
         reset_password_sent_at: パスワードリセット送信時刻
         reset_password_token: パスワードリセット用トークン
-        sign_in_count: ログイン回数
-        unconfirmed_email: 未確認Eメール
-        unlock_token: ロック解除用トークン
-        updated_at: 更新日
-  Edit_title: "%{name}編集"
-  Are_you_sure?: 本当によろしいですか？
-  Update: 更新
-  Cancel_my_account: アカウント削除
-  Sign_up: アカウント登録
-  Login: ログイン
-  Edit_profile: プロフィール編集
-  User_index: ユーザー一覧
-  Logout: ログアウト
-  Show: 詳細
-  Edit: 編集
-  Destroy: 削除
-  views.common.show: 詳細
-  Back: 戻る
-  logged_in_as: "%{login_email} ログイン中"
+    errors:
+      models:
+        user:
+          attributes:
+            email: 
+              taken: "はすでに使用されています"
+              blank: "は必須です"
+            password_confirmation:
+              confirmation: "がパスワードと一致していません"
+            password: 
+              too_short: "が短すぎます"
+            current_password:
+              blank: "が空欄です"
+              invalid: "が違います"
+  users: 
+    edit:
+      are_you_sure: 本当によろしいですか？
+      cancel_my_account: アカウント削除
+      currently_waiting_confirmation_for_email: "%{email} の確認待ち"
+      leave_blank_if_you_don_t_want_to_change_it: 空欄のままなら変更しません
+      title: "%{resource}編集"
+      unhappy: 気に入りません
+      update: 更新
+      we_need_your_current_password_to_confirm_your_changes: 変更を反映するには現在のパスワードを入力してください
+  controllers:
+    common:
+      notice_create: "新規作成しました"
+      notice_update: "本の情報を更新しました"
+      notice_destroy: "本を削除しました"
+  edit_title: "%{name}編集"
+  are_you_sure?: 本当によろしいですか？
+  update: 更新
+  cancel_my_account: アカウント削除
+  sign_up: アカウント登録
+  login: ログイン
+  edit_profile: プロフィール編集
+  user_index: ユーザー一覧
+  logout: ログアウト
   show: 詳細
   edit: 編集
   destroy: 削除
-  Book_index: 本一覧
+  views.common.show: 詳細
+  back: 戻る
+  logged_in_as: "%{login_email} ログイン中"
+  book_index: 本一覧
   new: 新規作成
-  Log_in: ログイン
+  log_in: ログイン
   title_new: 本の新規作成
   title_edit: "%{name}編集"
-  Change_your_password: パスワード変更
-  New_password: 新しいパスワード
-  Confirm_new_password: 新しいパスワード(確認)
-  Change_my_password: パスワード変更
+  change_your_password: パスワード変更
+  new_password: 新しいパスワード
+  confirm_new_password: 新しいパスワード(確認)
+  change_my_password: パスワード変更
   helpers:
     submit:
       Cancel_my_account: アカウントを削除する
@@ -63,14 +85,6 @@ ja:
       update: 更新する
     models:
       user: ユーザー
-    page_entries_info:
-      one_page:
-        display_entries:
-          zero: ""
-          one: "<strong>1-1</strong>/1件中"
-          other: "<strong>1-%{count}</strong>/%{count}件中"
-      more_pages:
-        display_entries: "<strong>%{first}-%{last}</strong>/%{total}件中"
   devise:
     confirmations:
       confirmed: メールアドレスが確認できました。
@@ -185,16 +199,9 @@ ja:
       not_saved:
         one: エラーが発生したため %{resource} は保存されませんでした。
         other: "%{count} 件のエラーが発生したため %{resource} は保存されませんでした。"
-    models:
-      user:
-        attributes:
-          password:
-            blank: "が空になっています"
-            too_short: "が短すぎます"
-  views:
-    pagination:
-      first: "&laquo; 最初"
-      last: "最後 &raquo;"
-      previous: "&lsaquo; 前"
-      next: "次 &rsaquo;"
-      truncate: "..."
+  views: 
+    common:
+      back: "戻る"
+      show: "詳細"
+      edit: "編集"
+      new: "新規作成"
